requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0
pathlib2>=2.3.0
